Changed drive selection method
Rather than fully wiping and re-partitioning the selected disk, the
script now accepts specific partitions for EFI, swap and the OS to
install to.

As part of this, if the EFI partition isn't formatted, it is done so
with FAT32 and a bootloader is installed.

If the swap partition is blank, it is made into a swap and mounted.

Multiple installs of Arch on the same system/disk is now supported, with
the hostname being used to split the boot images on the EFI partition.