Merge 'latest' versions into scripts
For Windows install scripts, the idea was to have a single central file
to update for version data, and calling that before calling individual
install scripts.

However, it's just easier in use to just have these versions embedded in
the install scripts, as calling the latest_versions.ps1 first is not
first or second nature.