Added ProgressPreference for powershell scripts
When downloading content via powershell's `Invoke-WebRequest`, by
default it will update a progress message after each incoming packet is
processed. However, this can largely extend the time it takes to
complete the request, such as 136MB file taking 5+ minutes.

Changing the `ProgressPreference` to be silent removes this overhead and
can massively reduce the time it takes to complete, such as the same
136MB file down to 30 seconds.