Update pre-commit copyright git script
The script would check against the working copy of the file instead of
what was staged. Thus, if the unstaged file has the correct year the
script would not raise an issue.

This fixes, to check agains tthe index version of the file, and will
only not throw a message if the staged file also has the correct year in
the header.